refactor: numbergenerator 패키지 domain 패키지 밖으로 분리
